+++ feature/google/src/main/java/com/revenuecat/purchases/google/BillingWrapper.kt
@@ -53,7 +53,7 @@ import com.revenuecat.purchases.strings.BillingStrings
 import com.revenuecat.purchases.strings.OfferingStrings
 import com.revenuecat.purchases.strings.PurchaseStrings
 import com.revenuecat.purchases.strings.RestoreStrings
-import com.revenuecat.purchases.utils.Result
+import com.revenuecat.purchases.Result
 import java.io.PrintWriter
 import java.io.StringWriter
 import java.util.Date
